+++ day_19/README.md
@@ -4,10 +4,13 @@
 
 `cat` みたいな子
 
-### FATに対応
+### 3. FATに対応
 
 FAT: File Allocate Table
 
 ファイルをどこに配置したのか的なやつ．
 説明をみてもなにもわからない...あとWindowsじゃない...
 
+### 4. ファイルの整理
+
+はい．
